+++ tests/place_holder.cpp
@@ -23,7 +23,7 @@ GTEST_TEST(PlaceHolder, PostPlaceHolder) {
   auto y = coll::place_holder<int>()
     | coll::map(anony_cc(_ + 1));
   auto z = coll::range(10) | y | x;
-  EXPECT_EQ(z.size(), 10);
+  EXPECT_EQ((int) z.size(), 10);
   for (int i = 0; i < 10; i++) {
     EXPECT_EQ(i + 1, z[i]);
   }
@@ -37,7 +37,7 @@ GTEST_TEST(PlaceHolder, PostPlaceHolder2) {
   auto w = coll::place_holder<int>()
     | coll::map(anony_cc(_));
   auto z = coll::range(10) | w | y | x;
-  EXPECT_EQ(z.size(), 10);
+  EXPECT_EQ((int) z.size(), 10);
   for (int i = 0; i < 10; i++) {
     EXPECT_EQ(i + 1, z[i]);
   }
@@ -58,19 +58,19 @@ GTEST_TEST(PlaceHolder, Concat) {
   z.start();
 
   z.run(coll::Left::value, 1);
-  EXPECT_EQ(v.size(), 1);
-  EXPECT_EQ(v.back(), 1 + 1);
+  EXPECT_EQ((int) v.size(), 1);
+  EXPECT_EQ((int) v.back(), 1 + 1);
 
   z.run(coll::Right::value, 2);
-  EXPECT_EQ(v.size(), 2);
+  EXPECT_EQ((int) v.size(), 2);
   EXPECT_EQ(v.back(), 2 * 2);
 
   z.run(coll::Left::value, 3);
-  EXPECT_EQ(v.size(), 3);
+  EXPECT_EQ((int) v.size(), 3);
   EXPECT_EQ(v.back(), 3 + 1);
 
   z.run(coll::Right::value, 4);
-  EXPECT_EQ(v.size(), 4);
+  EXPECT_EQ((int) v.size(), 4);
   EXPECT_EQ(v.back(), 4 * 2);
 
   z.end();
@@ -99,19 +99,19 @@ GTEST_TEST(PlaceHolder, Concat2) {
   w.start();
 
   w.run(coll::Left::value, coll::Left::value, 1);
-  EXPECT_EQ(v.size(), 1);
+  EXPECT_EQ((int) v.size(), 1);
   EXPECT_EQ(v.back(), 1 + 1);
 
   w.run(coll::Left::value, coll::Right::value, 2);
-  EXPECT_EQ(v.size(), 2);
+  EXPECT_EQ((int) v.size(), 2);
   EXPECT_EQ(v.back(), 2 * 2);
 
   w.run(coll::Right::value, coll::Left::value, 3);
-  EXPECT_EQ(v.size(), 3);
+  EXPECT_EQ((int) v.size(), 3);
   EXPECT_EQ(v.back(), 3 - 3);
 
   w.run(coll::Right::value, coll::Right::value, 4);
-  EXPECT_EQ(v.size(), 4);
+  EXPECT_EQ((int) v.size(), 4);
   EXPECT_EQ(v.back(), 4 / 4);
 
   w.end();
